# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Recipes API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  email: string;
  name?: string;
  picture?: string;
  isAdmin?: boolean;
}

export interface GoogleSignInRequest {
  credential: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface CategoryListResponse {
  categories: string[];
}

export interface RecipeSummary {
  id: string;
  slug: string;
  name: string;
  category: string;
  favorited?: boolean;
}

export interface RecipeSummaryListResponse {
  recipes: RecipeSummary[];
}

export interface ShortcutsResponse {
  recent: RecipeSummary[];
  random: RecipeSummary[];
}

export interface Ingredient {
  id: string;
  amountText: string;
  amountValue?: number;
  unit?: string;
  product: string;
  notes: string;
}

export type RecipeDetail = RecipeSummary & {
  yieldText: string;
  yieldServings?: number;
  ingredients: Ingredient[];
  steps: string[];
  utensils: string[];
};

export interface RecipeDetailResponse {
  recipe: RecipeDetail;
}

export interface CategoryRenameInput {
  name: string;
}

export interface IngredientDraft {
  amountText: string;
  amountValue?: number;
  unit?: string;
  product: string;
  notes: string;
}

export interface RecipeDraft {
  name: string;
  category: string;
  yieldText: string;
  yieldServings?: number;
  ingredients: IngredientDraft[];
  steps: string[];
}

export interface ParseRecipeInput {
  text: string;
}

export interface ParseRecipeResponse {
  recipe: RecipeDraft;
}

export type CreateRecipeInput = RecipeDraft & {
  utensils?: string[];
};

export interface SuccessResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
}

export type SearchRecipesParams = {
q?: string;
category?: string;
};