# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Recipes API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: recipes
    description: Recipe browsing, search, and detail operations
  - name: favorites
    description: Per-user favorite recipe operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/google:
    post:
      operationId: signInWithGoogle
      tags: [auth]
      summary: Sign in with Google
      description: Verifies a Google ID token and returns an app session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleSignInRequest"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the current signed-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /categories:
    get:
      operationId: listCategories
      tags: [recipes]
      summary: List distinct recipe categories, for the search-page filter chips
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryListResponse"
  /recipes/shortcuts:
    get:
      operationId: getSearchShortcuts
      tags: [recipes]
      summary: The signed-in user's last 3 searched recipes, plus 5 random recipes for discovery
      responses:
        "200":
          description: Shortcuts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShortcutsResponse"
  /recipes:
    get:
      operationId: searchRecipes
      tags: [recipes]
      summary: Search recipes by name or category
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Matching recipes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeSummaryListResponse"
    post:
      operationId: createRecipe
      tags: [recipes]
      summary: Create a new recipe from structured data (manually entered, or pre-filled by parseRecipeText)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecipeInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeDetailResponse"
  /recipes/parse:
    post:
      operationId: parseRecipeText
      tags: [recipes]
      summary: Use AI to extract structured recipe data (name, category, ingredients, steps) from pasted free-form recipe text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParseRecipeInput"
      responses:
        "200":
          description: Parsed recipe draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParseRecipeResponse"
        "502":
          description: The AI parser failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /recipes/{slug}:
    get:
      operationId: getRecipe
      tags: [recipes]
      summary: Get the full detail (ingredients, steps, utensils) for one recipe
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Recipe detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeDetailResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /recipes/{recipeId}:
    put:
      operationId: updateRecipe
      tags: [recipes]
      summary: Update a recipe's details, ingredients, steps, and utensils (admin only)
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecipeInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeDetailResponse"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /categories/{category}:
    put:
      operationId: renameCategory
      tags: [recipes]
      summary: Rename a category, updating every recipe currently tagged with it
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryRenameInput"
      responses:
        "200":
          description: Renamed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /recipes/{recipeId}/delete:
    delete:
      operationId: deleteRecipe
      tags: [recipes]
      summary: Permanently delete a recipe
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /favorites:
    get:
      operationId: listFavorites
      tags: [favorites]
      summary: List the signed-in user's favorited recipes
      responses:
        "200":
          description: Favorites
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeSummaryListResponse"
  /favorites/{recipeId}:
    put:
      operationId: addFavorite
      tags: [favorites]
      summary: Favorite a recipe
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Favorited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
    delete:
      operationId: removeFavorite
      tags: [favorites]
      summary: Unfavorite a recipe
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unfavorited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: "string"
        picture:
          type: "string"
        isAdmin:
          type: boolean
      required: [id, email]
    GoogleSignInRequest:
      type: object
      properties:
        credential:
          type: string
      required: [credential]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    CategoryListResponse:
      type: object
      required: [categories]
      properties:
        categories:
          type: array
          items:
            type: string
    RecipeSummary:
      type: object
      required: [id, slug, name, category]
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        category:
          type: string
        favorited:
          type: boolean
    RecipeSummaryListResponse:
      type: object
      required: [recipes]
      properties:
        recipes:
          type: array
          items:
            $ref: "#/components/schemas/RecipeSummary"
    ShortcutsResponse:
      type: object
      required: [recent, random]
      properties:
        recent:
          type: array
          items:
            $ref: "#/components/schemas/RecipeSummary"
        random:
          type: array
          items:
            $ref: "#/components/schemas/RecipeSummary"
    Ingredient:
      type: object
      required: [id, amountText, product, notes]
      properties:
        id:
          type: string
        amountText:
          type: string
        amountValue:
          type: number
        unit:
          type: string
        product:
          type: string
        notes:
          type: string
    RecipeDetail:
      allOf:
        - $ref: "#/components/schemas/RecipeSummary"
        - type: object
          required: [yieldText, ingredients, steps, utensils]
          properties:
            yieldText:
              type: string
            yieldServings:
              type: number
            ingredients:
              type: array
              items:
                $ref: "#/components/schemas/Ingredient"
            steps:
              type: array
              items:
                type: string
            utensils:
              type: array
              items:
                type: string
    RecipeDetailResponse:
      type: object
      required: [recipe]
      properties:
        recipe:
          $ref: "#/components/schemas/RecipeDetail"
    CategoryRenameInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
    IngredientDraft:
      type: object
      required: [amountText, product, notes]
      properties:
        amountText:
          type: string
        amountValue:
          type: number
        unit:
          type: string
        product:
          type: string
        notes:
          type: string
    RecipeDraft:
      type: object
      required: [name, category, yieldText, ingredients, steps]
      properties:
        name:
          type: string
        category:
          type: string
        yieldText:
          type: string
        yieldServings:
          type: number
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/IngredientDraft"
        steps:
          type: array
          items:
            type: string
    ParseRecipeInput:
      type: object
      required: [text]
      properties:
        text:
          type: string
    ParseRecipeResponse:
      type: object
      required: [recipe]
      properties:
        recipe:
          $ref: "#/components/schemas/RecipeDraft"
    CreateRecipeInput:
      allOf:
        - $ref: "#/components/schemas/RecipeDraft"
        - type: object
          properties:
            utensils:
              type: array
              items:
                type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string