# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Recipes API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  email: string;
  name?: string;
  picture?: string;
  isAdmin?: boolean;
}